modularize the search method dropdown for the filter form
